add Authorization

+++ App/controllers/ListingController.php
@@ -4,6 +4,8 @@ namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController {
     protected $db;
@@ -14,7 +16,7 @@ class ListingController {
     }
 
     public function index() {
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
         loadView('listings/index', [
             'listings' => $listings
@@ -54,7 +56,7 @@ loadView('listings/show', [
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
         
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -100,6 +102,8 @@ loadView('listings/show', [
 
         $this->db->query($query, $newListingData);
 
+        Session::setFlashMessage('success_message', 'Listing created successfully');
+
         redirect('/listings');
       }
     }
@@ -125,10 +129,16 @@ loadView('listings/show', [
                 return;
             }
 
+        // Authorization
+        if(!Authorization::isOwner($listing['user_id'])) {
+            Session::setFlashMessage('error_message', 'You are not authorized to delete this listing');
+            return redirect('/listings/' . $listing['id']);
+        }
+
         $this->db->query('DELETE from listings WHERE id = :id', $params);
 
         // Set flash message
-        $_SESSION['success_message'] = 'Listing deleted successfully';
+        Session::setFlashMessage('success_message', 'Listing deleted successfully');
 
         redirect('/listings');
      }
@@ -151,7 +161,12 @@ loadView('listings/show', [
                 return;
             }
 
-          
+               // Authorization
+               if(!Authorization::isOwner($listing['user_id'])) {
+                Session::setFlashMessage('error_message', 'You are not authorized to edit this listing');
+                return redirect('/listings/' . $listing['id']);
+            }
+    
 
             loadView('listings/edit', [
                 'listing' => $listing
@@ -178,6 +193,12 @@ loadView('listings/show', [
         if(!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+
+           // Authorization
+           if(!Authorization::isOwner($listing['user_id'])) {
+            Session::setFlashMessage('error_message', 'You are not authorized to update this listing');
+            return redirect('/listings/' . $listing['id']);
         }
 
         $allowedFields = ['title', 'description', 'salary', 'tags', 'company', 'address', 'city', 'state', 'phone', 'email', 'requirements', 'benefits'];
@@ -228,5 +249,30 @@ loadView('listings/show', [
        
       }
       
-      
+        /**
+      * Search listings by keywords/location
+      *
+      * @params array $params
+      * @return void
+      */
+
+      public function search() {
+        $keywords = isset($_GET['keywords']) ? trim($_GET['keywords']) : '';
+        $location= isset($_GET['location']) ? trim($_GET['location']) : '';
+
+        $query = "SELECT * FROM listings WHERE (title LIKE :keywords OR description LIKE :keywords OR tags LIKE :keywords OR company LIKE :keywords) AND (city LIKE :location OR state LIKE :location)";
+
+        $params = [
+            'keywords' => "%{$keywords}%",
+            'location' => "%{$location}%"
+        ];
+
+        $listings = $this->db->query($query, $params)->fetchAll();
+
+        loadView('/listings/index', [
+            'listings' => $listings,
+            'keywords' => $keywords,
+            'loaction' => $location
+        ]);
+      }
   }
